Change the structure a bit. index.ts only re-exports things. Move license. Specify correct license in package.json. Remove separate publish folder. Remove package script.
