+++ package.json
@@ -1,112 +1,51 @@
 {
-  
   "name": "@arction/lcjs-themes",
   "version": "1.0.0",
   "description": "Open-source Themes package for LightningChart JS",
   "keywords": [
-      "Arction",
-      "chart",
-      "data",
-      "lightning",
-      "visualization",
-      "xy",
-      "WebGL",
-      "graph",
-      "ohlc",
-      "candlestick",
-      "performance",
-      "spider",
-      "pie",
-      "funnel",
-      "linechart",
-      "gauges",
-      "trading",
-      "lightningchart",
-      "lightning chart",
-      "finance",
-      "plot",
-      "box plot",
-      "spline",
-      "multi-channel",
-      "scrolling chart",
-      "LineSeries",
-      "line series",
-      "line chart",
-      "visualisation",
-      "data visualisation",
-      "data visualization",
-      "bar chart",
-      "spline chart",
-      "high performance",
-      "dashboard",
-      "fast",
-      "big data",
-      "ekg",
-      "ecg",
-      "dataviz",
-      "GPU",
-      "typescript",
-      "javascript",
-      "charting library",
-      "charts",
-      "React",
-      "Vue",
-      "graphs",
-      "powerful",
-      "streaming",
-      "fastest",
-      "pyramid",
-      "heatmap",
-      "intensity",
-      "node js",
-      "zoom band",
-      "constant line",
-      "band",
-      "server-side rendering",
-      "spectrogram",
-      "3D",
-      "surface",
-      "box series",
-      "EEG",
-      "polar chart",
-      "logarithmic"
+    "Arction",
+    "lightningchart",
+    "lightning chart"
   ],
   "homepage": "https://github.com/Arction/lcjs-themes",
-  "license": "SEE LICENSE IN LICENSE",
+  "license": "MIT",
   "author": "LightningChart Ltd.",
   "files": [
-      "dist/lcjs-themes.d.ts",
-      "dist/lcjs-themes.iife.js",
-      "dist/lcjs-themes.js",
-      "dist/lcjs-themes.cjs",
-      "dist/lcjs-themes.mjs",
-      "dist/lcjs-themes.esm.js",
-      "dist/lcjs-themes.iife.es2019.js",
-      "LICENSE",
-      "package.json"
+    "dist/lcjs-themes.d.ts",
+    "dist/lcjs-themes.iife.js",
+    "dist/lcjs-themes.js",
+    "dist/lcjs-themes.cjs",
+    "dist/lcjs-themes.mjs",
+    "dist/lcjs-themes.esm.js",
+    "dist/lcjs-themes.iife.es2019.js",
+    "LICENSE",
+    "README.md",
+    "package.json"
   ],
   "main": "dist/lcjs-themes.js",
   "types": "dist/lcjs-themes.d.ts",
   "module": "dist/lcjs-themes.esm.js",
   "iife": "dist/lcjs-themes.iife.js",
   "exports": {
-      "import": "./dist/lcjs-themes.mjs",
-      "require": "./dist/lcjs-themes.cjs"
+    "import": "./dist/lcjs-themes.mjs",
+    "require": "./dist/lcjs-themes.cjs"
+  },
+  "engines": {
+    "node": ">=12"
   },
   "scripts": {
     "ts": "tsc -p ./tsconfig.json && tsc -p ./tsconfig.json --target ES5 --outDir temp/lcjs-themes/es5",
-    "build": "npm run ts && rollup -c rollup.config.js",
-    "package": "sh ./scripts/package.sh"
+    "build": "npm run ts && rollup -c rollup.config.js"
   },
   "dependencies": {
-    "@arction/lcjs": "file:v4.0.0-rc-2023-01-25"
+    "@arction/lcjs": "file:arction-lcjs-4.0.0.tgz"
   },
   "devDependencies": {
-    "rollup": "^2.77.0",
     "@rollup/plugin-babel": "^5.3.1",
     "@rollup/plugin-commonjs": "^22.0.2",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^14.0.0",
+    "rollup": "^2.77.0",
     "typescript": "4.8.4"
   }
 }
