--- publish/README.md
@@ -1,9 +1,0 @@
-# Open-source Themes package for LightningChart JS
-
-- Provides tools and ready-made options for customizing LightningChart JS colors and styles
-- Helps with kicking off your own custom themes development by checking through the open-source repository
-- No need to wait for new LightningChart JS versions to get updates for themes, colorization and styling 
-
-Learn more about the Themes package at [GitHub](https://github.com/Arction/lcjs-themes).
-
-Learn more about LightningChart JS [on our web site](https://lightningchart.com/js-charts/).
